Rearrange logic
