+++ casper-finality-proofs/src/test_engine/utils/mod.rs
@@ -1,3 +1,4 @@
 pub mod macros;
 pub mod parse_file;
+pub mod setup;
 pub mod test_engine;
